requirements.txt
aiohttp @ file:///croot/aiohttp_1670009560265/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
appdirs==1.4.4
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs @ file:///croot/attrs_1668696182826/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056706450/work
colorama @ file:///croot/colorama_1672386526460/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography @ file:///croot/cryptography_1677533068310/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
et-xmlfile==1.1.0
Flask @ file:///croot/flask_1671217343254/work
fonttools==4.25.0
frozenlist @ file:///croot/frozenlist_1670004507010/work
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna @ file:///croot/idna_1666125576474/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1666298844297/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib @ file:///croot/matplotlib-suite_1679593461707/work
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
multidict @ file:///croot/multidict_1665674239670/work
munkres==1.1.4
numexpr @ file:///croot/numexpr_1683221822650/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
openai @ file:///croot/openai_1682522392103/work
openpyxl==3.0.10
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
pandas-stubs @ file:///croot/pandas-stubs_1676319782725/work
Pillow==9.4.0
plotly @ file:///tmp/abs_7afcdfad-dbbb-49d2-adea-186abf525c45jbnd8p95/croots/recipe/plotly_1658160053621/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyPDF2==3.0.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytz @ file:///croot/pytz_1671697431263/work
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn @ file:///croot/scikit-learn_1680198742267/work
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tenacity @ file:///croot/tenacity_1682972284834/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tqdm @ file:///croot/tqdm_1679561862951/work
types-pytz @ file:///croot/types-pytz_1665514243479/work
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1680254681959/work
Werkzeug @ file:///croot/werkzeug_1679489717957/work
yarl @ file:///opt/conda/conda-bld/yarl_1661437085904/work
